fix a bug introduced in 31042b8d
  glDrawElements call expect unsigned int!

trivial bug-fix for master

+++ Polyhedron/demo/Polyhedron/Scene_edit_polyhedron_item.cpp
@@ -60,16 +60,16 @@ Scene_edit_polyhedron_item::Scene_edit_polyhedron_item
   tris.resize(polyhedron()->size_of_facets()*3);
   counter = 0;
   for(Polyhedron::Facet_handle fb = polyhedron()->facets_begin(); fb != polyhedron()->facets_end(); ++fb, ++counter) {
-    tris[counter*3] = fb->halfedge()->vertex()->id();
-    tris[counter*3+1] = fb->halfedge()->next()->vertex()->id();
-    tris[counter*3+2] = fb->halfedge()->prev()->vertex()->id();
+    tris[counter*3] =  static_cast<unsigned int>(fb->halfedge()->vertex()->id());
+    tris[counter*3+1] = static_cast<unsigned int>(fb->halfedge()->next()->vertex()->id());
+    tris[counter*3+2] = static_cast<unsigned int>(fb->halfedge()->prev()->vertex()->id());
   }
 
   edges.resize(polyhedron()->size_of_halfedges());
   counter = 0;
   for(Polyhedron::Edge_iterator eb = polyhedron()->edges_begin(); eb != polyhedron()->edges_end(); ++eb, ++counter) {
-    edges[counter*2] = eb->vertex()->id();
-    edges[counter*2+1] = eb->opposite()->vertex()->id();
+    edges[counter*2] = static_cast<unsigned int>(eb->vertex()->id());
+    edges[counter*2+1] = static_cast<unsigned int>(eb->opposite()->vertex()->id());
   }
 }
